requirements.txt
google-adk==1.21.0
requests>=2.31.0
